+++ src/modules/grid/grid_spline/Gridding_Spline_CSA.cpp
@@ -1,5 +1,5 @@
 /**********************************************************
- * Version $Id: Gridding_Spline_CSA.cpp 911 2011-02-14 16:38:15Z reklov_w $
+ * Version $Id: Gridding_Spline_CSA.cpp 1081 2011-06-08 08:05:26Z reklov_w $
  *********************************************************/
 
 ///////////////////////////////////////////////////////////
@@ -217,7 +217,7 @@ bool CGridding_Spline_CSA::On_Execute(void)
 	m_Points.Clear();
 
 	//-----------------------------------------------------
-	point	*pDst	= (point *)SG_Malloc(m_pGrid->Get_NCells() * sizeof(point));
+	point	*pDst	= (point *)SG_Malloc((long) m_pGrid->Get_NCells() * sizeof(point));
 
 	for(y=0, i=0, p.y=m_pGrid->Get_YMin(); y<m_pGrid->Get_NY() && Set_Progress(y, m_pGrid->Get_NY()); y++, p.y+=m_pGrid->Get_Cellsize())
 	{
