+++ src/modules_images/modules_imagery/imagery_segmentation/watershed_segmentation.cpp
@@ -1,5 +1,5 @@
 /**********************************************************
- * Version $Id: watershed_segmentation.cpp 911 2011-02-14 16:38:15Z reklov_w $
+ * Version $Id: watershed_segmentation.cpp 1081 2011-06-08 08:05:26Z reklov_w $
  *********************************************************/
 
 ///////////////////////////////////////////////////////////
@@ -335,7 +335,7 @@ bool CWatershed_Segmentation::Get_Segments(void)
 	int		Join		= Threshold > 0.0 ? Parameters("JOIN")->asInt() : 0;
 
 	//-----------------------------------------------------
-	for(int n=0; n<Get_NCells() && Set_Progress_NCells(n); n++)	
+	for(long n=0; n<Get_NCells() && Set_Progress_NCells(n); n++)	
 	{
 		int		x, y, i, ID, iID;
