+++ src/modules_simulation/hydrology/sim_hydrology/topmodel_values.cpp
@@ -1,5 +1,5 @@
 /**********************************************************
- * Version $Id: topmodel_values.cpp 911 2011-02-14 16:38:15Z reklov_w $
+ * Version $Id: topmodel_values.cpp 1084 2011-06-08 08:10:42Z reklov_w $
  *********************************************************/
 
 ///////////////////////////////////////////////////////////
@@ -169,7 +169,7 @@ void CTOPMODEL_Values::Destroy(void)
 //---------------------------------------------------------
 void CTOPMODEL_Values::Create(double dTime, int anTimeSteps, CSG_Parameters *pParameters, CSG_Grid *pAtanB, int anClasses, CSG_Grid *pClass)
 {
-	int		n, iClass, nCells;
+	long	n, iClass, nCells;
 
 	double	zMin, zRange, dz;
